simplify inigo server
don't actually need to NetIn since we always reach it via the container IP.

this happened to work before because we used NetIn with zero-value requested
host/port, where it'll pick a random one and use it for both. what we
technically should have been using was the container port, not the host port;
and only used NetIn if we wanted to reach this via the VM's external address.