use LONG_TIMEOUT in eventually, 10 in consistently
for consistently, the value given is time for assurance, not time after which
to give up. as we bump the 'give up' timeout (i.e. for testing on systems
under load), the consistency checks should not take any longer.

in this case, it's testing that things don't happen concurrently, and so the
next assertion should have happened around the same amount of time.

Signed-off-by: Abhijit Hiremagalur <abhi@pivotallabs.com>